JCBC-815: Do not let the CouchbaseEnvironment.Builder implement the Env IF.
Motivation
----------
A constant source of confusion has been the fact that the CouchbaseEnvironment
Builder also implements the CouchbaseEnvironment. Also, this meant that methods
which are not supported on the builder needed to throw exception or (worse)
return null.

Modifications
-------------
This changeset removes the interface implementation and consequently gets
rid of the unneeded getters on the builder.

The environment code has been adapted to access the fields in the builder
directly, instead through its getter methods.

Result
------
The builder is now clearly distinguishable from the built environment
itself and also does not have to throw unsupported exceptions on methods
which are not meant to be called at all.

Change-Id: I30b9d10214244c10c25c1a85e14c3ae6ffb8dfbf
Reviewed-on: http://review.couchbase.org/53543
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>