JCBC-817: Rename parametrized to parameterized for consistency reasons.
Motivation
----------
It was decided, while parametrized is correct in form, that throughout
documentation and SDKs the term "parameterized" should be used instead,
since it appears to be the primary form according to the oxford
dictionary.

Since this API hasn't been stableized yet until 2.2.0, we can make this
change now better than later.

Modifications
-------------
This changeset replaces all parametrized occurences with the parameterized
form.

Result
------
Consistent naming of concepts across docs and SDKs.

Change-Id: I1dc030bedf687a77bad2c54b0ce887c1cdc39e53
Reviewed-on: http://review.couchbase.org/53736
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>