JCBC-819: Implement a Bucket Consumer for Events
Motivation
----------
It makes sense to allow storing events in a couchbase bucket,
mainly to showcase how it works. In production it might come
in helpful if those events are sent to a different cluster
which collects and aggregates those statistics.

Modifications
-------------
Adding the BucketConsumer and allow to use either insert or
update of the documents.

Result
------
A simple to plug consumer which writes the events into a
bucket as json data.

Change-Id: Id9f3a9bba1721f468a110d5d3160f2cb7077d25b
Reviewed-on: http://review.couchbase.org/53798
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>