JCBC-823: Expose maxParallelism as QueryParam.
Motivation
----------
For performance optimizations on queries, the maxParallelism
param has been added to optionally override the server side
default.

Modifications
-------------
The QueryParam now exposes the option and passes it down to
the query engine if set. Tests adapted to make sure it works
as specified.

One oddity is that the server expects the number as a string,
so our code converts the int to a string.

Result
------
MaxParallelism supported for n1ql queries.

Change-Id: I42985f7e8ca5396b4f76148b04f67f439c2317d5
Reviewed-on: http://review.couchbase.org/53931
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>