JCBC-825: Expose MutationToken on Document
Motivation
----------
As a prepration for enhanced durability and optimized n1ql queries, the
mutation token needs to be exposed on the document itself.

Modifications
-------------
The Document API has been extended to support the token, but since it is optional
it is backwards compatible (nulling out if not supported).

If the user does not want to use mutation tokens, even custom documents will continue
to work, only if mutation tokens are enabled the custom documents need to implement
one more method to create the document with the mutation token.

Result
------
The mutation token is exposed and can be utilized from user and sdk space.

Change-Id: Id7bf5eb5c3e67f027f746ba92dd15f9c73eea9c2
Reviewed-on: http://review.couchbase.org/54133
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>