JCBC-828: Always apply a server side timeout on query.
Motivation
----------
In order to avoid queries running forever on the server side,
the client side timeout should be always be applied, since
there is no point in running the server side execution code
when the client has timed out and is not listening anymore.

Modifiations
------------
By default, the 75s environment timeout will be used on the
async API, as well as the sync API. If the sync api overrides
the client side timeout, it will be mirrored. Note that
the user can always tune the timeout on the QueryParams,which
always takes precedence. This is important for super long running
queries which might take longer than 75s.

Result
------
Avoid running queries infinitely on the server if the client
stopped waiting some time ago.

Change-Id: I434dd438f304ca71ca5cee27e01b21fde15b44a3
Reviewed-on: http://review.couchbase.org/54168
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>