JCBC-829: Change API for new prepared statements handling
Motivation
----------
The agreed spec includes transparently handling the prepared statement,
so a new adhoc query option needs to be introduced, as well as the
prepare() methods removed from the public API.

Modifications
-------------
The prepare() API is removed from the async and sync API, and the new
query option is added to the query params.

Result
------
Cleanup for upcoming prepared statement changes.

Change-Id: I673d7c66efb454cba412f24fbff36ad01e3f5d76
Reviewed-on: http://review.couchbase.org/54410
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>