JCBC-829: Add LRU Query Cache
Motivation
----------
Since prepared objects are nor transparently cached, we need a thread
safe LRU cache to handle the objects.

Modifications
-------------
Without pulling in a nother library, for now we synchronize a
LinkedHashMap extension which provides LRU semantics with very
little code necessary. In the future, custom implementation can
provide better throughput, but it should be more than enough
to start with.

Result
------
LRU cache implemented, ready to be used in later code.

Change-Id: I7457549c2168f552a18778e9967d5240a50fcfcc
Reviewed-on: http://review.couchbase.org/54412
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>