JCBC-829: Internalize Prepared Statements Execution
Motivation
----------
The final step of the prepared statement rework includes using
the LRU cache and conditionally either executing the cached
prepared payload or running the prepare code first.

Modifications
-------------
The code and tests have been modified to remove all external
visibility of "prepared objects", and also now the cache wipe
method is exposed on the bucket API.

Result
------
Automatic and internal execution of prepared statements, but
without retry logic on errors (which will come in a later
commit).

Change-Id: Iaf97e9a08e60a168db75c01f7b875ca3485ade29
Reviewed-on: http://review.couchbase.org/54444
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>