JCBC-830: Handle EXISTS response on append/prepend.
Motivation
----------
When an append/prepend operation is called with the cas
value set and the requests value does not match with the
server one, it will return with an EEXISTS (same as on
a replace operation, for example).

Modifications
-------------
The proper mapping from EXISTS to a CASMismatchException
has been coded, as well as adding integration tests to
avoid regressions in the future.

Result
------
EXISTS on append/prepend are now properly converted into
the right exception type.

Change-Id: I7a054503a8bf1dbbbfdb76e5d7d8c8fc4fb2da1a
Reviewed-on: http://review.couchbase.org/54498
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>