JCBC-836: Durability Requirements for append, prepend and counter.
Motivation
----------
In addition to the regular mutation methods, append, prepend and counter
return cas values (or mutation tokens) so in turn they can be observed
for durability requirements.

The proper overloads need to be exposed for public consumption.

Modifications
-------------
Thanks to no default arguments in java, a plethora of overloads
has been added to both the sync and async API to expose durability
requirements on the append, prepend and counter methods.

Result
------
Support for durability requirements on append, prepend and counter.

Change-Id: Iee1777f8da4d4958ed9787741e75b5465559aaf5
Reviewed-on: http://review.couchbase.org/54810
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>