JCBC-829: Evict from cache before retry and add tests
When a prepared payload is scheduled for retry, first evict it from
the cache (so that even if re-prepare fails, we'll have a chance to
try again next time the same statement is issued).

Added tests on the extraction of the plan, content of the execute
prepared query (server name, along with "encoded_plan" for
the compressed plan...).

Change-Id: I76bd50b04f4ac80ebc9b667425519af2495bdba5
Reviewed-on: http://review.couchbase.org/54851
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>