Added mocked unit tests that experiment with back-pressure
First a classical bulk get pattern scenario is demonstrated, where
BackpressureException is thrown because the ringbuffer is filled (as
simulated by the mock reacting to queue reaching MAX_CAPACITY).

Secondly, a new pattern is demonstrated where natural flow control in
RxJava is used to avoid that BackpressureException, by using the flatMap
maxConcurrent variant.

Change-Id: Ia54b4cbf0a41b4017c45dc904df59fd904df394d
Reviewed-on: http://review.couchbase.org/55489
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>