JCBC-852: Improve docs on view management
Motivation:
-----------
Documentation on view management is a bit unclear, especially on how to
add/update a specific view into an existing DesignDocument (what is the
behavior of upsert in this case? can one use getDesignDocument, mutate
the views() list in it and then upsert?).

Modifications:
--------------
Added javadoc on various view-related elements.
Documented how to add a view on upsertDesignDocument's javadoc.
Documented reduce nullability.
Documented ways to create a concrete instance of View.

Added integration tests that showcase and test the behavior of upsert
DesignDocument when one doesn't resend the whole list of view definition
vs doing a get/add view/upsert of the DesignDocument.

This shows that adding a new definition of an existing view will update
it on the server side.

Results:
--------
View management is better documented in javadocs and an integration test
verifies the behavior when adding a view.

Change-Id: I1124b87cad3b1adcde0d164d6dc588ade22a9a00
Reviewed-on: http://review.couchbase.org/55726
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>