JCBC-850: Initial API for Full Text.
Motivation
----------
The Couchbase Server "Watson" release will add support for
full text indexing. As a result, the SDKs need to provide
first-class APIs to make it possible to easily query those
indexes.

Modifications
-------------
This changeset introduces a highly experimental API for
CBFT that is subject to change, but will provide users
a straightforward way to try it out and give feedback
in the developer preview and beta timeframe.

Result
------
First experimental CBFT API.

Change-Id: I601be9ad35abb2966fe38dc1a2dce413c9c94345
Reviewed-on: http://review.couchbase.org/55845
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>