JCBC-856: Update README for 2.2.1
Motivation
----------
The README has been outdated after the change from
gradle back to maven. Also, this is a great opportunity
for more clarity.

Also, a CONTRIBUTING file with some basic information
is helpful for newcomers.

Modifications
-------------
The README is now restructured and clarified, also
providing new information on how to contribute.

The CONTRIBUTING file has been added with similar
info than in the README.

Result
------
Easier to digest README with updated infos.

Change-Id: If52b3fe7134b9a4e917df84689f1b75d36a5c097
Reviewed-on: http://review.couchbase.org/55989
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>