JCBC-863: Cache Bucket and AsyncBucket references.
Motivation
----------
Best practice recommends that only one Cluster and one Bucket
reference per bucket are opened and then reused as a singleton.

In the field it has been discovered that sometimes the buckets
are not reused, but rather opened on every call and just the
Cluster object is reused.

Modifications
-------------
To apply "defense in depth" if the best practices are not
followed, the Bucket and AsyncBucket objects are cashed in
a CHM so when subsequent calls are incoming, no new Bucket
instances are created.

Lifecyclemanagement is also includes, so when a bucket is
closed and then opened again, this is discovered and a
the references are replaced.

Result
------
Even when the best practices are not followed, this change
tries to minimize the impact.

Change-Id: Ib4835643fac1a9b56f10d6ffe725b5afb1f0c843
Reviewed-on: http://review.couchbase.org/56153
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>