JCBC-777: Index Management
Motivation:
-----------
Secondary index management (listing, creating and dropping indexes on a
particular bucket) is something that can be considered a repetitive
task for some categories of users and as such can be offered as an API
on top of the corresponding N1QL requests.

Modifications:
--------------
Introduced several methods in the BucketManager and AsyncBucketManager
to deal with secondary indexes. This API only deals with GSI indexes.

Methods to watch specific indexes, polling until they become online (in
the case of a deferred build) are also part of the API.

Modified the IndexNamesPath to be able to trigger a build on a List of
indexes (rather than having to rework such a list into a first parameter
followed by a vararg).

Added integration tests for the API.

Results:
--------
Users can now manage (GSI) indexes more easily than by crafting adhoc
N1QL queries.

They can poll for indexes to become online and see the polling progress
in logs by activating DEBUG on the "indexWatch" named Logger.

Change-Id: I9b4224bf7b640e4e2543a53a5747ba418e6f8a75
Reviewed-on: http://review.couchbase.org/57530
Well-Formed: buildbot <build@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>