JCBC-873: Point out that mutating cluster ops are async
Motivation
----------
The previous javadocs did not make it clear that even if  the server
responds with a success and the observable/call is completed, the
operation on the server itself is asynchronous and might not be
finished yet.

Modifications
-------------
Enhance javadoc to point out the fact.

Result
------
More clarity in the javadoc on server side behaviour for cluster
management operations.

Change-Id: I1cc3e0884ae4f99eb8db24a93ad805abbd879f9b
Reviewed-on: http://review.couchbase.org/56625
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>