JCBC-872: Add Javadocs for parameterized queries
Motivation
----------
A quick sample on how to use positional or named parameterized
queries needs to be part of the javadocs to ease adoption.

Modifications
-------------
The N1qlQuery methods have been extended with javadoc blocks
explaining the concepts and showing examples on how to use them.

Result
------
Helping with easier adoption for parameterized queries.

Change-Id: I761657d092b43713d05e0f30aba0c1ccb7c8fab1
Reviewed-on: http://review.couchbase.org/56629
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>