JCBC-844: Clarify javadocs of upsert.
Motivation
----------
It has been brought to attention that the javadoc for upsert is
misleading and does not clarify that upsert is not actually performing
a replace operation and does not take the CAS into account.

Modifications
-------------
The javadocs now use overwrite instead of replace in the wording,
as well as contain a paragraph which states that CAS is not respected
and the corresponding replace method needs to be used instead.

Result
------
More clarity for the upsert javadocs.

Change-Id: Ic7a3aaf7d96661e9348857f8e5785c25be05d2f9
Reviewed-on: http://review.couchbase.org/56630
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>