JCBC-871: N1ql status exposed in result
Motivation:
-----------
So far only the boolean status of a N1QL response was exposed on the
result object. This has become kind of confusing when users started to
use server-side timeouts, since in case of timeout only the status is
changed to "timeout" (no errors, no rows). So it makes sense to expose
the raw status as well.

Modifications:
--------------
N1QL status is exposed through the status() method. The finalSuccess()
method now relies on "status()" in the DefaultAsyncN1qlQueryResult.

Result:
-------
Exact status of a N1QL query can always be determined by using status().

Change-Id: I9e1053be04493d7932b9e4efb77a1691c279c76b
Reviewed-on: http://review.couchbase.org/56642
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>