JCBC-876: Add integration tests for DML
Motivation
----------
Since 4.0 ships DML as experimental and 4.1 will ship it as GA, it
makes sense for the java client to have appropriate tests in place.

Modifications
-------------
This changeset starts with this effort and adds integration tests
for INSERT, UPSERT, UPDATE and DELETE. One test for DELETE has
been ignored for now until MB-16732 is resolved.

Note that MERGE tests have not been added yet. More tests will
certainly be added as the development of the fluent API for those
kinds of statements progresses.

Result
------
First batch of integration tests for N1QL DML.

Change-Id: I90b2fa24ad692c53bf7199f1841815cdd849b169
Reviewed-on: http://review.couchbase.org/56738
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>