JCBC-881: Expose custom transcoder with durability requirements.
Motivation
----------
On the CouchbaseClientIF, all the mutation methods expose a custom
transcoder, but only without durability requirements. It is a valid
use case to use a custom transcoder with durability requirements.

Modifications
-------------
Add API which allows to use the durability requirements together
with a custom transcoder.

The changeset itself is low risk, since the methods are available
internally, only the approrpiate overloads have been added which
now call the underlying methods explicitly with the transcoder.

Result
------
Mutations with a custom transcoder and durability requirements
can now be used easily.

Change-Id: I5eca2162d3b5c844635dbf1b8d0aa7c03edcdb8f
Reviewed-on: http://review.couchbase.org/57103
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>