JCBC-888: Improve Maven-based javadoc generation
Motivation:
-----------
The Maven-based build had a few missing elements not ported from the
Gradle build (footer, title, etc...).
Additionally, the java-client javadoc didn't link to the core-io javadoc
(eg for classes like CouchbaseEnvironment having parent classes in the
core).

Modifications:
--------------
All elements of the javadoc previously set in the Gradle build are also
set in the Maven build.

Added offline linking of the core-io javadoc. This is done by locally
unpacking the core-io javadoc jar and using that as the target for
linking (since the docs are not yet published on the web at this point).
The generated urls reflect the url where the core docs are expected to
be published.

Results:
--------
Better javadoc packaged by the Maven build.

Change-Id: I1057bd1c05d9696bcc579613a3e1a701d011b824
Reviewed-on: http://review.couchbase.org/57462
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>