JCBC-804: Implement basic AT_PLUS support.
Motivation
----------
Couchbase Server Watson adds an extended feature set for
optimized "Read Your Own Wrties" queries that in some use
cases performs better than plain queries with REQUEST_PLUS
consistency.

Specifically, providing the mutation results which contains
the mutation token allows to grab a lower bound for sequence
numbers on the server side, avoiding to wait until the indexer
has caught up with the latest seqno on all vbuckets.

Modifications
-------------
This changeset implements the majority use case for AT_PLUS
support as specified by the SDK-RFC spec (no 4).

It implements it for both Document and DocumentFragment when
it is passed in with the MutationToken set. Both other use
cases (Bucket only as well as only doc id) warrant larger
changes in the core and will be implemented in separate
changesets.

Result
------
Optimized RYOW support based on AT_PLUS and mutation tokens
is usable.

Change-Id: Iedc14746dcc5c254c6fdc1845c423e6dc719cbe0
Reviewed-on: http://review.couchbase.org/57844
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>