JCBC-883: Add fromCurrentBucket() to select builder API.
Motivation
----------
Most of the time when using the select builder API, one needs to
write select(..).from(i("bucket-name")), where the bucket name
reflects the current bucket in scope.

We can make this easier for the user.

Modifications
-------------
The builder adds a fromCurrentBucket() method which internally
adds a #CURRENT_BUCKET# identifier into the n1ql string which
is replaced with the bucket name as `` before it is send to
the core engine. That way, the builder API can make it easier
to just say "load the data from the current bucket".

Result
------
Easier access to the current bucket in the builder API.

Change-Id: I9d1fffe590d31c2b61bd726133ee92f48d1cecf2
Reviewed-on: http://review.couchbase.org/58040
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>