JCBC-849: Subdocument client API
Motivation:
-----------
Subdocument API allows to mutate and read specific fragments of JSON
inside an existing document without having to transfer the whole
document over the wire.

Modifications:
--------------
Added the DocumentFragment class, both serving as an input parameter
for mutation and a return parameter. Additionally to what would be
available in a top-level Document, it also bears a path (pointing to a
specific value inside the JSON).

Added Bucket level API for Subdoc:
 - simple operations (getIn, existIn, insertIn, replaceIn, upsertIn,
   removeIn)
 - array operations (extendIn, arrayInsertIn, addUniqueIn)
 - arithmetic operation (counterIn)
 - multiple lookups (lookupIn)
 - multiple mutations (mutateIn)

Multi operations have their own return types.

Added associated exceptions as subclasses of SubDocumentException.

A dedicated interface for transcoding in subdoc (FragmentTranscoder)
has been added. A default implementation, JacksonFragmentTranscoder, is
based on Jackson and shares code with JsonTranscoder, so this code was
mutualized into TranscoderUtils.
This default cannot be customized for now but could be in the future.

Result:
-------
The subdoc API can be used for efficient modification of JSON documents.
Integration tests validate the behavior of the API on cluster >= 4.5.0.

Change-Id: Ifce305713329f0d63073fab3012129211e4c41f6
Reviewed-on: http://review.couchbase.org/58223
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>