JCBC-838: ViewQuery toString is for logging, added toQueryString
Motivation:
-----------
toString shouldn't usually have special semantics that would prevent
changing the representation (as it is more for human consumption, logs,
etc...).

Currently ViewQuery.toString outputs the parameters that would go into
a GET url, omitting isDevelopment, isIncludeDocs and keys.

Modifications:
--------------
toString is now purely about string representation, including the design
document name, the view name, "dev" if isDevelopment == true,
"includeDocs" if isIncludeDocs == true, as well as the content of keys
if filled in.

Keys dump will be truncated to 140 characters (since it can grow quite
large in the body) and information about the total length is also
displayed if truncated.

The former value of toString is now returned by "toQueryString()", which
is used as the GET parameters payload when executing the query.

Result:
-------
toString now displays more information and can be further evolved.
toQueryString is used for payload.

Change-Id: Ic30ff590d60a5008fcd4905b4b8526917568529d
Reviewed-on: http://review.couchbase.org/59015
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>