JCBC-909: Add toString for LookupSpec/MutationSpec
Motivation
----------
When getting results from a multi-operation in subdoc, users will
probably want to output the specification next to the result at some
point. This is not made easy by the fact that the specs have no proper
toString method.

Modifications
-------------
Added short toString implementation for both multi specs.
Format is "{COMMAND:path}".

MutationSpec doesn't output the fragment value, as it could be large,
but suffixes the COMMAND with ", createParents" if this option is
activated on the mutation.

Result
------
A LookupSpec or a MutationSpec can be used in an output now that their
toString method is implemented.

Change-Id: I3b2c97a4ef2906412686a36bf2ddd0c4cfd2ba00
Reviewed-on: http://review.couchbase.org/59133
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>