Disable Subdoc multi-mutate support
Motivation:
-----------
The protocol for multiple mutations is likely to change before a DP
version of the server is released, so offering support for it is too
early.

Modifications:
--------------
Made classes and code related to multi mutation requests/responses not
public and/or commented. The API mutateIn has been commented out.

Result:
-------
Single subdoc operations and multi lookups are still available as
experimental, but multi-mutation is not.

Change-Id: I7c37b885ac16ca584c5a2629a681c25230119992
Reviewed-on: http://review.couchbase.org/59251
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>