JCBC-912: Explicitly check that the retry information is not null.
Motivation
----------
If the server returns a NMVB with an empty response, the code needs
to check for this case and do not potentially throw a NPE if this
case remains undetected.

Modifications
-------------
The code checks if the buffer which potentially contains the new
config is not null and if it is just returns (nothing to do here
at this point).

Result
------
Better resiliency for the NMVB config optimizations implemented
in the server for the watson timeframe.

Change-Id: Ie236ed0beb26251c4f9d84ef911906e508b0d41a
Reviewed-on: http://review.couchbase.org/59258
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>