JVMCBC-286: Enforce usage of Java 6 APIs only
Motivation:
-----------
When developing against a recent JDK, it's not sufficient to just set
the targetVersion to 1.6, as the compiler will happily produce Java 6
bytecode that uses APIs introduced in a more recent version (eg Java 8).
This results in code that can't run on a Java 6 JRE.

The Animal-Sniffer tool has a Maven integration that allows to check for
such bad usage and prevent the build process to finish.

Modifications:
--------------
Used Animal-Sniffer Maven plugin to enforce correct usage of JDK 6 APIs
only, executed during compile phase.

Result:
-------
Using a too new API when developing against a recent JDK will be
detected when using Maven to build, and refused.

Change-Id: I4fca83105e8cdd9a5c0e8031b20543711a437221
Reviewed-on: http://review.couchbase.org/59342
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>