JCBC-804: Bring AT_PLUS in line with the latest RFC changes.
Motivation
----------
The latest RFC simplifies the proposal by reducing some of the
feature set for the first iteration. Since those concepts were
not implemented yet, the actual java implementation also got
simplified and streamlined.

Modifications
-------------
Method signatures updated and internal data structures and
processing simplified as well.

Note that the unique seqno match has been updated and is now
done when the tokens are inserted, rather than when it is
processed. This also optimizes for the case where only ever
on token is passed in.

Result
------
Java code in line with the latest RFC iteration.

Change-Id: Ie74d3b812c72ffb2fee982879b6325964ded5f3f
Reviewed-on: http://review.couchbase.org/60243
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>