JCBC-875: retrybuilder advanced error predicate + doOnRetry
Motivation:
-----------
When using the RetryBuilder, one can have an even more complex check to
perform on the exception do determine if it should be retried or not.
Currently the only possible check is on the exception type, but not its
content (eg. message, or special fields in the exception).

Similarly, sometimes one wants to perform a side-effect action when the
retry occurs, and this is not possible in the current implementation.

Modification:
-------------
Added two features to the RetryBuilder:

 * one can construct the builder based on a Func1<Throwable, Boolean>
   predicate, that will allow to tell if the error should trigger a
   retry, allowing inspection of the whole error.
 * added a doOnRetry() method that gets notified every time a retry is
   scheduled with: the retry attempt number, the exception triggering
   the retry, the scheduled delay (duration and unit) after which the
   retry will occur.

The signature for the doOnRetry, a specialized Func4, can be simplified
using the RetryBuilder.OnRetryAction interface.

Results:
--------
RetryBuilder gives even more control on when to retry, and allows you to
add side-effects to each retry attempt.

Change-Id: Idecf6403a5bfa5896a682422a2636c10d0fa3e2d
Reviewed-on: http://review.couchbase.org/60562
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>